Move special slide creation to slidefactory
