+++ slideshow/src/main/java/nl/ou/services/AbstractSlideFactory.java
@@ -1,8 +1,11 @@
 package nl.ou.services;
 
+import java.util.List;
+
 import nl.ou.domain.Content;
 import nl.ou.domain.Slide;
 import nl.ou.domain.SlideMeta;
+import nl.ou.domain.SlideshowMeta;
 import nl.ou.services.impl.SimpleSlideFactory;
 
 public abstract class AbstractSlideFactory {
@@ -11,6 +14,10 @@ public abstract class AbstractSlideFactory {
     public abstract SlideMeta createSlideMeta(String title, int number);
 
     public abstract Slide createSlide(SlideMeta meta, Content content);
+
+    public abstract Slide createTOCSlide(List<SlideMeta> metas);
+
+    public abstract Slide createTitleSlide(SlideshowMeta meta);
 
     public static AbstractSlideFactory getSlideFactory() {
         if (SLIDE_FACTORY == null) {
