+++ slideshow/src/main/java/nl/ou/services/impl/SimpleSlideFactory.java
@@ -1,8 +1,12 @@
 package nl.ou.services.impl;
 
+import java.util.List;
+
 import nl.ou.domain.*;
 import nl.ou.domain.impl.NormalSlide;
 import nl.ou.domain.impl.SimpleSlideMeta;
+import nl.ou.domain.impl.TitleSlide;
+import nl.ou.domain.impl.TocSlide;
 import nl.ou.services.AbstractSlideFactory;
 
 public class SimpleSlideFactory extends AbstractSlideFactory {
@@ -14,5 +18,15 @@ public class SimpleSlideFactory extends AbstractSlideFactory {
     @Override
     public Slide createSlide(SlideMeta meta, Content content) {
         return new NormalSlide(content, meta);
+    }
+
+    @Override
+    public Slide createTOCSlide(List<SlideMeta> metas) {
+        return new TocSlide(metas);
+    }
+
+    @Override
+    public Slide createTitleSlide(SlideshowMeta meta) {
+        return new TitleSlide(meta);
     }
 }
